+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-05 13:22:49
-// Commit: refactor(core): extract reusable atomic widgets from core
+// Updated at: 2026-04-06 11:03:51
+// Commit: chore(core): update dependencies and cleanup unused imports in core
 
 class coreComponentState {
-  final String commitRef = '140';
-  final String lastMessage = 'refactor(core): extract reusable atomic widgets from core';
-  final DateTime timestamp = DateTime.parse('2026-04-05T13:22:49+05:00');
+  final String commitRef = '150';
+  final String lastMessage = 'chore(core): update dependencies and cleanup unused imports in core';
+  final DateTime timestamp = DateTime.parse('2026-04-06T11:03:51+05:00');
 }